+++ src/core/urls.py
@@ -16,10 +16,11 @@ Including another URLconf
 """
 
 from django.conf.urls.i18n import i18n_patterns
+from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import include, path
 
-from .views import favicon_view
+from core import settings
 
 urlpatterns = (
     *i18n_patterns(
@@ -27,5 +28,7 @@ urlpatterns = (
         path("", include("home.urls")),
         path("i18n/", include("django.conf.urls.i18n")),
     ),
-    path("favicon.ico", favicon_view, name="favicon"),
 )
+
+if settings.DEBUG:
+    urlpatterns = (*urlpatterns, *static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT))