+++ src/home/urls.py
@@ -1,8 +1,5 @@
 from django.urls import path
 
-from home.views import background_view, home_view
+from home.views import home_view
 
-urlpatterns = (
-    path("", home_view),
-    path("background.jpg", background_view, name="background"),
-)
+urlpatterns = (path("", home_view),)