+++ src/home/views.py
@@ -1,7 +1,5 @@
-from django.core.files.storage import default_storage
 from django.http import HttpRequest, HttpResponse
 from django.shortcuts import render
-from django.views.decorators.cache import cache_page
 
 from .models import Experience, PersonalInfo
 
@@ -19,9 +17,3 @@ def home_view(request: HttpRequest) -> HttpResponse:
             ),
         },
     )
-
-
-@cache_page(60 * 60 * 24)
-def background_view(request: HttpRequest) -> HttpResponse:
-    with default_storage.open("background.jpg", "rb") as favicon:
-        return HttpResponse(favicon.read(), content_type="image/jpeg")